requirements.txt
pandas
requests
beautifulsoup4
sentence_transformers
faiss-cpu
openai==0.28
